# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API каталога исполнителей и их альбомов с песнями",
        "description": "API каталога исполнителей и их альбомов с песнями",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/songs/": {
            "get": {
                "tags": [
                    "Песни"
                ],
                "summary": "Получение информации о песнях",
                "description": "Получение информации о песнях",
                "operationId": "GetSongs",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SongResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Песни"
                ],
                "summary": "Добавление песни",
                "description": "Добавление песни",
                "operationId": "PostSong",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SongStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SongResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Сообщение ошибки валидатора",
                                            "type": "string",
                                            "example": "Поле 'name' должно быть строкой"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/songs/{song_id}": {
            "patch": {
                "tags": [
                    "Песни"
                ],
                "summary": "Обновление информации о песне",
                "description": "Обновление информации о песне",
                "operationId": "PatchSong",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SongUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SongResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Entity Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Сообщение ошибки валидатора",
                                            "type": "string",
                                            "example": "Поле 'name' должно быть строкой"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/songs/{songs_id}": {
            "delete": {
                "tags": [
                    "Песни"
                ],
                "summary": "Удаление песни",
                "description": "Удаление песни",
                "operationId": "DeleteSong",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Исполнитель не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Entity Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SongStoreRequest": {
                "description": "Данные запроса для добавления пени",
                "required": [
                    "name",
                    "orderNumberInAlbum",
                    "albumId"
                ],
                "properties": {
                    "name": {
                        "description": "Название песни",
                        "type": "string",
                        "example": "I Am Hated"
                    },
                    "orderNumberInAlbum": {
                        "description": "Порядковый номер в альбоме",
                        "type": "integer",
                        "example": "10"
                    },
                    "albumId": {
                        "description": "Id альбома",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
            "SongUpdateRequest": {
                "description": "Данные запроса для обновления данных альбома",
                "properties": {
                    "name": {
                        "description": "Название песни",
                        "type": "string",
                        "example": "Снаружи всех измерени"
                    },
                    "orderNumberInAlbum": {
                        "description": "Порядковый номер в альбоме",
                        "type": "integer",
                        "example": "2"
                    },
                    "albumId": {
                        "description": "Id альбома",
                        "type": "integer",
                        "example": "16"
                    }
                },
                "type": "object"
            },
...
            "SongResource": {
                "description": "Данные о песне",
                "properties": {
                    "id": {
                        "description": "Id песни",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "Название песни",
                        "type": "string",
                        "example": "Wait and Bleed"
                    },
                    "orderNumberInAlbum": {
                        "description": "Порядковый номер в альбоме",
                        "type": "integer",
                        "example": "4"
                    },
                    "album": {
                        "description": "Информация об альбоме",
                        "properties": {
                            "id": {
                                "description": "Id альбома",
                                "type": "integer",
                                "example": "2"
                            },
                            "name": {
                                "description": "Название альбома",
                                "type": "string",
                                "example": "Iowa"
                            },
                            "yearOfIssue": {
                                "description": "Год выпуска",
                                "type": "integer",
                                "example": "2001"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
...
    }
}